// apps/template/src/api/generated/types.gen.ts
export type VoloAbpPermissionManagementPermissionGrantInfoDto = {
    name?: string | null;
    displayName?: string | null;
    parentName?: string | null;
    isGranted?: boolean;
    allowedProviders?: Array<string> | null;
    grantedProviders?: Array<VoloAbpPermissionManagementProviderInfoDto> | null;
    isEditable?: boolean;
};

export type VoloAbpPermissionManagementPermissionGroupDto = {
    name?: string | null;
    displayName?: string | null;
    displayNameKey?: string | null;
    displayNameResource?: string | null;
    permissions?: Array<VoloAbpPermissionManagementPermissionGrantInfoDto> | null;
};

export type VoloAbpPermissionManagementProviderInfoDto = {
    providerName?: string | null;
    providerKey?: string | null;
};